# Improve performance of node selector matcher