… staging/src/k8s.io/apimachinery/pkg/labels/zz_generated.deepcopy.go