@yuzhiquan 2022-02-17
scheduling benchmarks on my laptop

before
```
/private/var/folders/jj/hmhqkjy11f7d0jltb6mpr8c00000gn/T/GoLand/___BenchmarkFilter_in_k8s_io_kubernetes_pkg_scheduler_framework_plugins_podtopologyspread.test -test.v -test.paniconexit0 -test.bench ^\QBenchmarkFilter\E$ -test.run ^$
goos: darwin
goarch: amd64
pkg: k8s.io/kubernetes/pkg/scheduler/framework/plugins/podtopologyspread
cpu: Intel(R) Core(TM) i5-1038NG7 CPU @ 2.00GHz
BenchmarkFilter
BenchmarkFilter/1000nodes/single-constraint-zone
BenchmarkFilter/1000nodes/single-constraint-zone-8         	    1759	    750968 ns/op
BenchmarkFilter/1000nodes/single-constraint-zone/Clone
BenchmarkFilter/1000nodes/single-constraint-zone/Clone-8   	  752032	      1662 ns/op
BenchmarkFilter/1000nodes/single-constraint-node
BenchmarkFilter/1000nodes/single-constraint-node-8         	    1263	    953243 ns/op
BenchmarkFilter/1000nodes/single-constraint-node/Clone
BenchmarkFilter/1000nodes/single-constraint-node/Clone-8   	   12343	     98375 ns/op
BenchmarkFilter/1000nodes/two-Constraints-zone-node
BenchmarkFilter/1000nodes/two-Constraints-zone-node-8      	     921	   1236147 ns/op
BenchmarkFilter/1000nodes/two-Constraints-zone-node/Clone
BenchmarkFilter/1000nodes/two-Constraints-zone-node/Clone-8         	   12152	     96975 ns/op
PASS
```
after
```
/private/var/folders/jj/hmhqkjy11f7d0jltb6mpr8c00000gn/T/GoLand/___BenchmarkFilter_in_k8s_io_kubernetes_pkg_scheduler_framework_plugins_podtopologyspread.test -test.v -test.paniconexit0 -test.bench ^\QBenchmarkFilter\E$ -test.run ^$
goos: darwin
goarch: amd64
pkg: k8s.io/kubernetes/pkg/scheduler/framework/plugins/podtopologyspread
cpu: Intel(R) Core(TM) i5-1038NG7 CPU @ 2.00GHz
BenchmarkFilter
BenchmarkFilter/1000nodes/single-constraint-zone
BenchmarkFilter/1000nodes/single-constraint-zone-8         	    1766	    680068 ns/op
BenchmarkFilter/1000nodes/single-constraint-zone/Clone
BenchmarkFilter/1000nodes/single-constraint-zone/Clone-8   	  772818	      1655 ns/op
BenchmarkFilter/1000nodes/single-constraint-node
BenchmarkFilter/1000nodes/single-constraint-node-8         	    1450	    741098 ns/op
BenchmarkFilter/1000nodes/single-constraint-node/Clone
BenchmarkFilter/1000nodes/single-constraint-node/Clone-8   	   12687	     95398 ns/op
BenchmarkFilter/1000nodes/two-Constraints-zone-node
BenchmarkFilter/1000nodes/two-Constraints-zone-node-8      	     979	   1213292 ns/op
BenchmarkFilter/1000nodes/two-Constraints-zone-node/Clone
BenchmarkFilter/1000nodes/two-Constraints-zone-node/Clone-8         	   12525	     95070 ns/op
PASS
```