@yuzhiquan 2022-02-18
> A few more considerations:
> 
> Since benchmarks can be tricky to gauge, I benchmarked over 10 and 50 runs and used [`benchstat`](https://pkg.go.dev/golang.org/x/perf/cmd/benchstat) to compare the performance before this change and after this change. The command I used to run the benchmark is:
> 
> ```
> go test -run XXX -bench ^BenchmarkFilter$ -count 10 >> ~/new_sched_10.txt
> ```
> 
> and similarly for the benchmark before this change. On comparing the results, we get the following
> 
> For 50 runs:
> 
> ```
> benchstat -sort delta ~/old_sched.txt ~/new_sched.txt
> name                                                old time/op  new time/op   delta
> Filter/1000nodes/single-constraint-node/Clone-8     95.0µs ± 1%  100.4µs ±15%   +5.70%  (p=0.000 n=41+47)
> Filter/1000nodes/two-Constraints-zone-node-8        1.38ms ± 8%   1.43ms ±23%   +3.95%  (p=0.001 n=46+44)
> Filter/1000nodes/two-Constraints-zone-node/Clone-8  96.8µs ± 2%   98.1µs ± 2%   +1.37%  (p=0.000 n=36+41)
> Filter/1000nodes/single-constraint-zone/Clone-8     1.66µs ± 7%   1.62µs ±11%     ~     (p=0.964 n=30+50)
> Filter/1000nodes/single-constraint-node-8           1.14ms ±17%   1.24ms ±48%     ~     (p=0.198 n=48+50)
> Filter/1000nodes/single-constraint-zone-8           1.42ms ±26%   0.97ms ± 2%  -32.11%  (p=0.000 n=49+46)
> ```
> 
> For 10 runs:
> 
> ```
> benchstat -sort delta ~/old_sched_10.txt ~/new_sched_10.txt
> name                                                old time/op  new time/op  delta
> Filter/1000nodes/single-constraint-zone-8           1.06ms ± 5%  1.19ms ±11%  +12.27%  (p=0.000 n=9+10)
> Filter/1000nodes/two-Constraints-zone-node-8        1.39ms ± 2%  1.48ms ±16%   +6.07%  (p=0.019 n=10+10)
> Filter/1000nodes/single-constraint-zone/Clone-8     1.73µs ±17%  1.54µs ± 4%     ~     (p=0.052 n=10+9)
> Filter/1000nodes/single-constraint-node-8           1.19ms ± 5%  1.30ms ±13%     ~     (p=0.052 n=10+10)
> Filter/1000nodes/single-constraint-node/Clone-8     88.8µs ±24%  93.1µs ±11%     ~     (p=0.095 n=9+10)
> Filter/1000nodes/two-Constraints-zone-node/Clone-8  98.3µs ± 2%  97.9µs ± 3%     ~     (p=0.645 n=8+8)
> ```
> 
> Additionally, we also have more allocations happening after the change is made (my guess is because of the global map of operators acting as a cache?) Old (this is just for one run of the benchmark to get the allocs):
> 
> ```
> goos: linux
> goarch: amd64
> pkg: k8s.io/kubernetes/pkg/scheduler/framework/plugins/podtopologyspread
> cpu: Intel(R) Core(TM) i7-8550U CPU @ 1.80GHz
> BenchmarkFilter/1000nodes/single-constraint-zone-8         	    1237	   1015809 ns/op	  413811 B/op	    2065 allocs/op
> BenchmarkFilter/1000nodes/single-constraint-zone/Clone-8   	  492946	      2353 ns/op	    1495 B/op	       9 allocs/op
> BenchmarkFilter/1000nodes/single-constraint-node-8         	    1064	   1096974 ns/op	  511108 B/op	    2065 allocs/op
> BenchmarkFilter/1000nodes/single-constraint-node/Clone-8   	    8775	    133160 ns/op	   99112 B/op	      10 allocs/op
> BenchmarkFilter/1000nodes/two-Constraints-zone-node-8      	     885	   1367161 ns/op	  673509 B/op	    4576 allocs/op
> BenchmarkFilter/1000nodes/two-Constraints-zone-node/Clone-8         	    9243	    112293 ns/op	   99160 B/op	      11 allocs/op
> ```
> 
> New:
> 
> ```
> goos: linux
> goarch: amd64
> pkg: k8s.io/kubernetes/pkg/scheduler/framework/plugins/podtopologyspread
> cpu: Intel(R) Core(TM) i7-8550U CPU @ 1.80GHz
> BenchmarkFilter/1000nodes/single-constraint-zone-8         	    1257	    972454 ns/op	  413979 B/op	    2067 allocs/op
> BenchmarkFilter/1000nodes/single-constraint-zone/Clone-8   	  510312	      1960 ns/op	    1495 B/op	       9 allocs/op
> BenchmarkFilter/1000nodes/single-constraint-node-8         	    1135	   1218474 ns/op	  511116 B/op	    2067 allocs/op
> BenchmarkFilter/1000nodes/single-constraint-node/Clone-8   	   12579	     95917 ns/op	   99112 B/op	      10 allocs/op
> BenchmarkFilter/1000nodes/two-Constraints-zone-node-8      	     854	   1356366 ns/op	  672331 B/op	    4580 allocs/op
> BenchmarkFilter/1000nodes/two-Constraints-zone-node/Clone-8         	   12871	     90156 ns/op	   99160 B/op	      11 allocs/op
> ```
> 
> Currently, I'm not sure if the speedups are significant enough to warrant additional allocs (which again aren't all that much). Very possible I'm missing something here, please lmk if that's the case.
> 
> It also might be worth exploring what the results are by increasing the benchtime.

That's impressive!thanks for the test, i will give another try!