@yuzhiquan 2022-02-18
> Replacing switch for maps is not very likely to give meaningful speedups.
> 
> Also, what I want to avoid is the switch during `Requirement.Matches` as that's the one that we want to use repeatedly.
> 
> Not sure how to explain my idea better: when creating a `Requirement` (in `NewRequirement`), we know what the operation is, so we can store a matcher `func` in the `Requirement` that we can directly call, instead of storing the operation `string` and doing the switch again.
> 
> An alternative approach is that `Requirement` is an interface and there are different implementations of it, but I think this is overkill.

I see, i will give another try soon!