@yuzhiquan 2022-03-04
> benchstat

in this commit, we get this benchmark:
```shell
> /Users/yuzhiquan/go/bin/benchstat old_50.txt new_50.txt
name                                                old time/op  new time/op  delta
Filter/1000nodes/single-constraint-zone-8            847µs ±44%   774µs ±78%  -8.58%  (p=0.001 n=49+45)
Filter/1000nodes/single-constraint-zone/Clone-8     1.62µs ± 5%  1.61µs ± 9%    ~     (p=0.124 n=45+50)
Filter/1000nodes/single-constraint-node-8            878µs ±22%   886µs ±39%    ~     (p=0.638 n=50+50)
Filter/1000nodes/single-constraint-node/Clone-8     90.3µs ± 5%  89.2µs ± 5%    ~     (p=0.085 n=49+43)
Filter/1000nodes/two-Constraints-zone-node-8        1.15ms ±16%  1.14ms ±16%    ~     (p=0.159 n=48+48)
Filter/1000nodes/two-Constraints-zone-node/Clone-8  89.6µs ± 8%  91.8µs ± 9%  +2.48%  (p=0.001 n=46+45)
```