@yuzhiquan 2022-03-07
New benchmark results:
```shell
> /Users/yuzhiquan/go/bin/benchstat old_50.txt new_50-2.txt
name                                                old time/op  new time/op  delta
Filter/1000nodes/single-constraint-zone-8            847µs ±44%   773µs ±32%    ~     (p=0.052 n=49+44)
Filter/1000nodes/single-constraint-zone/Clone-8     1.62µs ± 5%  1.69µs ±14%  +4.26%  (p=0.000 n=45+48)
Filter/1000nodes/single-constraint-node-8            878µs ±22%   924µs ±29%    ~     (p=0.076 n=50+49)
Filter/1000nodes/single-constraint-node/Clone-8     90.3µs ± 5%  91.8µs ±10%  +1.65%  (p=0.031 n=49+46)
Filter/1000nodes/two-Constraints-zone-node-8        1.15ms ±16%  1.19ms ±11%  +3.28%  (p=0.002 n=48+42)
Filter/1000nodes/two-Constraints-zone-node/Clone-8  89.6µs ± 8%  89.6µs ± 7%    ~     (p=0.215 n=46+43)
```