@yuzhiquan 2022-03-07
New benchmark result:
```shell
> /Users/yuzhiquan/go/bin/benchstat old_50-2.txt new_50-2.txt
name                                                old time/op  new time/op  delta
Filter/1000nodes/single-constraint-zone-8            793µs ±36%   773µs ±32%    ~     (p=0.777 n=50+44)
Filter/1000nodes/single-constraint-zone/Clone-8     1.58µs ± 4%  1.69µs ±14%  +6.56%  (p=0.000 n=46+48)
Filter/1000nodes/single-constraint-node-8            897µs ±30%   924µs ±29%    ~     (p=0.289 n=47+49)
Filter/1000nodes/single-constraint-node/Clone-8     95.5µs ±20%  91.8µs ±10%  -3.84%  (p=0.007 n=47+46)
Filter/1000nodes/two-Constraints-zone-node-8        1.24ms ±10%  1.19ms ±11%  -3.46%  (p=0.000 n=46+42)
Filter/1000nodes/two-Constraints-zone-node/Clone-8  96.3µs ±21%  89.6µs ± 7%  -6.96%  (p=0.000 n=49+43
```