@yuzhiquan 2022-03-07
New benchmark result:
```shell
> /Users/yuzhiquan/go/bin/benchstat old_50-3.txt new_50-3.txt
name                                                old time/op  new time/op  delta
Filter/1000nodes/single-constraint-zone-8            886µs ±31%   751µs ±33%  -15.28%  (p=0.000 n=49+50)
Filter/1000nodes/single-constraint-zone/Clone-8     1.61µs ± 2%  1.54µs ± 3%   -4.25%  (p=0.000 n=47+42)
Filter/1000nodes/single-constraint-node-8            867µs ±19%   815µs ±31%   -6.06%  (p=0.000 n=46+50)
Filter/1000nodes/single-constraint-node/Clone-8     91.8µs ± 4%  88.3µs ± 2%   -3.77%  (p=0.000 n=45+44)
Filter/1000nodes/two-Constraints-zone-node-8        1.21ms ±12%  1.25ms ±34%     ~     (p=0.870 n=49+50)
Filter/1000nodes/two-Constraints-zone-node/Clone-8  90.1µs ± 2%  87.7µs ± 3%   -2.71%  (p=0.000 n=46+46)
```