@yuzhiquan 2022-03-09
Benchmark result:
```shell
> /Users/yuzhiquan/go/bin/benchstat old_50-4.txt new_50-4.txt
name                                                old time/op  new time/op  delta
Filter/1000nodes/single-constraint-zone-8            852µs ±47%   687µs ±18%  -19.39%  (p=0.000 n=50+47)
Filter/1000nodes/single-constraint-zone/Clone-8     1.64µs ±11%  1.56µs ± 4%   -5.15%  (p=0.000 n=44+45)
Filter/1000nodes/single-constraint-node-8            885µs ±40%   756µs ±18%  -14.55%  (p=0.000 n=44+49)
Filter/1000nodes/single-constraint-node/Clone-8     90.1µs ±11%  88.5µs ± 3%   -1.75%  (p=0.040 n=45+47)
Filter/1000nodes/two-Constraints-zone-node-8        1.15ms ±10%  1.15ms ±17%     ~     (p=0.511 n=49+47)
Filter/1000nodes/two-Constraints-zone-node/Clone-8  87.8µs ± 4%  87.0µs ± 2%   -0.92%  (p=0.000 n=43+47)
```